requirements.txt
playwright==1.41.2
asyncio==3.4.3
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.1
aiohttp==3.9.3
pytest==8.0.1
httpx==0.27.0
requests==2.31.0